moved code from Portfolio.js to App.js
